requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Science and Numerical Computing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning and Evaluation
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0

# For reproducibility
# Note: Install PyTorch with appropriate CUDA version if GPU is available
# Visit: https://pytorch.org/get-started/locally/
